Add files via upload
This is a mitigation of a security vulnerability in haveibeenpwend code that I did for my Introduction to Secure Software class. It is based on a fictional Cable TV company called InstaNeTV. The vulnerability lets the users create official company accounts for themselves, using their personal email addresses, regardless of how many times their email accounts may have been breached in data spills. The mitigation that is applied to the code, searches the email addresses in haveibeenpwned server to find out home many times it may have been breached. If an email address is found to have been breached more than 3 times, it advices the users to use other email addresses instead. 
Accompanying the project code is a PDF document that discusses the mitigation process.  The API key in the code has been replaced with a fake key, based on our professor's request. To run the code, an active and valid API keys is required.